cleanup: Move GenericSymbolRef to SCIPSymbolRef.{h,cc}
