+++ scip_indexer/SCIPIndexer.cc
@@ -29,7 +29,6 @@
 #include "core/Loc.h"
 #include "core/SymbolRef.h"
 #include "core/Symbols.h"
-#include "main/lsp/lsp.h"
 #include "main/pipeline/semantic_extension/SemanticExtension.h"
 #include "sorbet_version/sorbet_version.h"
 
@@ -89,251 +88,6 @@ struct OwnedLocal {
         // 32-bits => if there are 10k methods in a single file, the chance of at least one
         // colliding pair is about 1.1%, assuming even distribution. That seems OK.
         return fmt::format("local {}~#{}", counter, ::fnv1a_32(owner.name(gs).show(gs)));
-    }
-};
-
-bool isSorbetInternal(const core::GlobalState &gs, core::SymbolRef sym) {
-    UnorderedSet<core::SymbolRef> visited;
-    auto classT = core::Symbols::T().data(gs)->lookupSingletonClass(gs);
-    while (sym.exists() && !visited.contains(sym)) {
-        if (sym.isClassOrModule()) {
-            auto klass = sym.asClassOrModuleRef();
-            if (klass == core::Symbols::Sorbet_Private() || klass == core::Symbols::T() || klass == classT) {
-                return true;
-            }
-            auto name = klass.data(gs)->name;
-            if (name == core::Names::Constants::Opus()) {
-                return true;
-            }
-        }
-        visited.insert(sym);
-        sym = sym.owner(gs);
-    }
-    return false;
-}
-
-template <typename T> using SmallVec = InlinedVector<T, 2>;
-
-// A wrapper type to handle both top-level symbols (like classes) as well as
-// "inner symbols" like fields (@x). In a statically typed language, field
-// symbols are like any other symbols, but in Ruby, they aren't (necessarily)
-// declared ahead-of-time (you can declare them with @x = T.let(…, …) though).
-// So Sorbet represents them with a separate name on the side.
-//
-// Structurally, this is similar to the Alias instruction. One key difference
-// is that a GenericSymbolRef may refer to the owner in some situations.
-class GenericSymbolRef final {
-    core::SymbolRef selfOrOwner;
-
-    /// Name of the symbol, which may or may not exist.
-    core::NameRef name;
-
-    /// The type of the symbol at its definition, if applicable.
-    ///
-    /// References to this symbol may have a different type,
-    /// because you can change the type of a field, including within
-    /// the same basic block.
-    core::TypePtr _definitionType;
-
-public:
-    enum class Kind {
-        ClassOrModule,
-        UndeclaredField,
-        DeclaredField,
-        Method,
-    };
-
-private:
-    GenericSymbolRef(core::SymbolRef s, core::NameRef n, core::TypePtr t, Kind k)
-        : selfOrOwner(s), name(n), _definitionType(t) {
-        switch (k) {
-            case Kind::ClassOrModule:
-                ENFORCE(s.isClassOrModule());
-                ENFORCE(!n.exists());
-                return;
-            case Kind::DeclaredField:
-                ENFORCE(s.isFieldOrStaticField());
-                ENFORCE(!n.exists());
-                return;
-            case Kind::UndeclaredField:
-                ENFORCE(s.isClassOrModule());
-                ENFORCE(n.exists());
-                return;
-            case Kind::Method:
-                ENFORCE(s.isMethod());
-                ENFORCE(!n.exists());
-        }
-    }
-
-public:
-    GenericSymbolRef(const GenericSymbolRef &) = default;
-    GenericSymbolRef(GenericSymbolRef &&) = default;
-    GenericSymbolRef &operator=(const GenericSymbolRef &) = default;
-    GenericSymbolRef &operator=(GenericSymbolRef &&) = default;
-
-    friend bool operator==(const GenericSymbolRef &lhs, const GenericSymbolRef &rhs) {
-        return lhs.selfOrOwner == rhs.selfOrOwner && lhs.name == rhs.name;
-    }
-
-    friend bool operator<(const GenericSymbolRef &lhs, const GenericSymbolRef &rhs) {
-        return lhs.selfOrOwner.rawId() < rhs.selfOrOwner.rawId() ||
-               (lhs.selfOrOwner == rhs.selfOrOwner && lhs.name.rawId() < rhs.name.rawId());
-    }
-
-    template <typename H> friend H AbslHashValue(H h, const GenericSymbolRef &c) {
-        return H::combine(std::move(h), c.selfOrOwner, c.name);
-    }
-
-    static GenericSymbolRef classOrModule(core::SymbolRef self) {
-        return GenericSymbolRef(self, {}, {}, Kind::ClassOrModule);
-    }
-
-    static GenericSymbolRef undeclaredField(core::SymbolRef owner, core::NameRef name, core::TypePtr type) {
-        return GenericSymbolRef(owner, name, type, Kind::UndeclaredField);
-    }
-
-    static GenericSymbolRef declaredField(core::SymbolRef self, core::TypePtr type) {
-        return GenericSymbolRef(self, {}, type, Kind::DeclaredField);
-    }
-
-    static GenericSymbolRef method(core::SymbolRef self) {
-        return GenericSymbolRef(self, {}, {}, Kind::Method);
-    }
-
-    core::TypePtr definitionType() const {
-        return this->_definitionType;
-    }
-
-    Kind kind() const {
-        if (this->name.exists()) {
-            return Kind::UndeclaredField;
-        }
-        if (this->selfOrOwner.isFieldOrStaticField()) {
-            return Kind::DeclaredField;
-        }
-        if (this->selfOrOwner.isMethod()) {
-            return Kind::Method;
-        }
-        return Kind::ClassOrModule;
-    }
-
-    UntypedGenericSymbolRef withoutType() const {
-        switch (this->kind()) {
-            case Kind::UndeclaredField:
-                ENFORCE(this->selfOrOwner.isClassOrModule());
-                return UntypedGenericSymbolRef::undeclared(this->selfOrOwner.asClassOrModuleRef(), this->name);
-            case Kind::Method:
-            case Kind::ClassOrModule:
-            case Kind::DeclaredField:
-                return UntypedGenericSymbolRef::declared(this->selfOrOwner);
-        }
-    }
-
-    /// Display a GenericSymbolRef for debugging.
-    string showRaw(const core::GlobalState &gs) const {
-        switch (this->kind()) {
-            case Kind::UndeclaredField:
-                return fmt::format("UndeclaredField(owner: {}, name: {})", this->selfOrOwner.showFullName(gs),
-                                   this->name.toString(gs));
-            case Kind::DeclaredField:
-                return fmt::format("DeclaredField {}", this->selfOrOwner.showFullName(gs));
-            case Kind::ClassOrModule:
-                return fmt::format("ClassOrModule {}", this->selfOrOwner.showFullName(gs));
-            case Kind::Method:
-                return fmt::format("Method {}", this->selfOrOwner.showFullName(gs));
-        }
-    }
-
-    core::SymbolRef asSymbolRef() const {
-        ENFORCE(this->kind() != Kind::UndeclaredField);
-        return this->selfOrOwner;
-    }
-
-    bool isSorbetInternalClassOrMethod(const core::GlobalState &gs) const {
-        switch (this->kind()) {
-            case Kind::UndeclaredField:
-            case Kind::DeclaredField:
-                return false;
-            case Kind::ClassOrModule:
-            case Kind::Method:
-                return isSorbetInternal(gs, this->asSymbolRef());
-        }
-    }
-
-    void saveDocStrings(const core::GlobalState &gs, core::TypePtr fieldType, core::Loc loc, SmallVec<string> &docs) {
-        auto checkType = [&gs, &loc](core::TypePtr ty, const std::string &name) {
-            ENFORCE(ty, "missing type for {} in file {}\n{}\n", name, loc.file().data(gs).path(), loc.toString(gs));
-        };
-
-        string markdown = "";
-        switch (this->kind()) {
-            case Kind::UndeclaredField: {
-                auto name = this->name.show(gs);
-                checkType(fieldType, name);
-                markdown = fmt::format("{} ({})", name, fieldType.show(gs));
-                break;
-            }
-            case Kind::DeclaredField: {
-                auto fieldRef = this->selfOrOwner.asFieldRef();
-                auto name = fieldRef.showFullName(gs);
-                checkType(fieldType, name);
-                markdown = fmt::format("{} ({})", name, fieldType.show(gs));
-                break;
-            }
-            case Kind::ClassOrModule: {
-                auto ref = this->selfOrOwner.asClassOrModuleRef();
-                auto classOrModule = ref.data(gs);
-                if (classOrModule->isClass()) {
-                    auto super = classOrModule->superClass();
-                    if (super.exists() && super != core::Symbols::Object()) {
-                        markdown = fmt::format("class {} < {}", ref.show(gs), super.show(gs));
-                    } else {
-                        markdown = fmt::format("class {}", ref.show(gs));
-                    }
-                } else {
-                    markdown = fmt::format("module {}", ref.show(gs));
-                }
-                break;
-            }
-            case Kind::Method: {
-                auto ref = this->selfOrOwner.asMethodRef();
-                auto resultType = ref.data(gs)->owner.data(gs)->resultType;
-                checkType(resultType, fmt::format("result type for {}", ref.showFullName(gs)));
-                markdown = realmain::lsp::prettyTypeForMethod(gs, ref, resultType, nullptr, nullptr);
-                // FIXME(varun): For some reason, it looks like a bunch of public methods
-                // get marked as private here. Avoid printing misleading info until we fix that.
-                // https://github.com/sourcegraph/scip-ruby/issues/33
-                markdown = absl::StrReplaceAll(markdown, {{"private def", "def"}, {"; end", ""}});
-                break;
-            }
-        }
-        if (!markdown.empty()) {
-            docs.push_back(fmt::format("```ruby\n{}\n```", markdown));
-        }
-        auto whatFile = loc.file();
-        if (whatFile.exists()) {
-            if (auto doc = realmain::lsp::findDocumentation(whatFile.data(gs).source(), loc.beginPos())) {
-                docs.push_back(doc.value());
-            }
-        }
-    }
-
-    core::Loc symbolLoc(const core::GlobalState &gs) const {
-        switch (this->kind()) {
-            case Kind::Method: {
-                auto method = this->selfOrOwner.asMethodRef().data(gs);
-                if (!method->nameLoc.exists() || method->nameLoc.empty()) {
-                    return method->loc();
-                }
-                return method->nameLoc;
-            }
-            case Kind::ClassOrModule:
-            case Kind::DeclaredField:
-                return this->selfOrOwner.loc(gs);
-            case Kind::UndeclaredField:
-                ENFORCE(false, "case UndeclaredField should not be triggered here");
-                return core::Loc();
-        }
     }
 };
