+++ scip_indexer/SCIPSymbolRef.cc
@@ -6,8 +6,10 @@
 #include <vector>
 
 #include "absl/status/status.h"
+#include "absl/strings/str_replace.h"
 
 #include "core/Loc.h"
+#include "main/lsp/lsp.h"
 
 #include "scip_indexer/SCIPSymbolRef.h"
 
@@ -78,4 +80,116 @@ absl::Status UntypedGenericSymbolRef::symbolForExpr(const core::GlobalState &gs,
     }
     return absl::OkStatus();
 }
+
+string GenericSymbolRef::showRaw(const core::GlobalState &gs) const {
+    switch (this->kind()) {
+        case Kind::UndeclaredField:
+            return fmt::format("UndeclaredField(owner: {}, name: {})", this->selfOrOwner.showFullName(gs),
+                               this->name.toString(gs));
+        case Kind::DeclaredField:
+            return fmt::format("DeclaredField {}", this->selfOrOwner.showFullName(gs));
+        case Kind::ClassOrModule:
+            return fmt::format("ClassOrModule {}", this->selfOrOwner.showFullName(gs));
+        case Kind::Method:
+            return fmt::format("Method {}", this->selfOrOwner.showFullName(gs));
+    }
+}
+
+bool GenericSymbolRef::isSorbetInternal(const core::GlobalState &gs, core::SymbolRef sym) {
+    UnorderedSet<core::SymbolRef> visited;
+    auto classT = core::Symbols::T().data(gs)->lookupSingletonClass(gs);
+    while (sym.exists() && !visited.contains(sym)) {
+        if (sym.isClassOrModule()) {
+            auto klass = sym.asClassOrModuleRef();
+            if (klass == core::Symbols::Sorbet_Private() || klass == core::Symbols::T() || klass == classT) {
+                return true;
+            }
+            auto name = klass.data(gs)->name;
+            if (name == core::Names::Constants::Opus()) {
+                return true;
+            }
+        }
+        visited.insert(sym);
+        sym = sym.owner(gs);
+    }
+    return false;
+}
+
+void GenericSymbolRef::saveDocStrings(const core::GlobalState &gs, core::TypePtr fieldType, core::Loc loc,
+                                      SmallVec<string> &docs) const {
+    auto checkType = [&gs, &loc](core::TypePtr ty, const std::string &name) {
+        ENFORCE(ty, "missing type for {} in file {}\n{}\n", name, loc.file().data(gs).path(), loc.toString(gs));
+    };
+
+    string markdown = "";
+    switch (this->kind()) {
+        case Kind::UndeclaredField: {
+            auto name = this->name.show(gs);
+            checkType(fieldType, name);
+            markdown = fmt::format("{} ({})", name, fieldType.show(gs));
+            break;
+        }
+        case Kind::DeclaredField: {
+            auto fieldRef = this->selfOrOwner.asFieldRef();
+            auto name = fieldRef.showFullName(gs);
+            checkType(fieldType, name);
+            markdown = fmt::format("{} ({})", name, fieldType.show(gs));
+            break;
+        }
+        case Kind::ClassOrModule: {
+            auto ref = this->selfOrOwner.asClassOrModuleRef();
+            auto classOrModule = ref.data(gs);
+            if (classOrModule->isClass()) {
+                auto super = classOrModule->superClass();
+                if (super.exists() && super != core::Symbols::Object()) {
+                    markdown = fmt::format("class {} < {}", ref.show(gs), super.show(gs));
+                } else {
+                    markdown = fmt::format("class {}", ref.show(gs));
+                }
+            } else {
+                markdown = fmt::format("module {}", ref.show(gs));
+            }
+            break;
+        }
+        case Kind::Method: {
+            auto ref = this->selfOrOwner.asMethodRef();
+            auto resultType = ref.data(gs)->owner.data(gs)->resultType;
+            checkType(resultType, fmt::format("result type for {}", ref.showFullName(gs)));
+            markdown = realmain::lsp::prettyTypeForMethod(gs, ref, resultType, nullptr, nullptr);
+            // FIXME(varun): For some reason, it looks like a bunch of public methods
+            // get marked as private here. Avoid printing misleading info until we fix that.
+            // https://github.com/sourcegraph/scip-ruby/issues/33
+            markdown = absl::StrReplaceAll(markdown, {{"private def", "def"}, {"; end", ""}});
+            break;
+        }
+    }
+    if (!markdown.empty()) {
+        docs.push_back(fmt::format("```ruby\n{}\n```", markdown));
+    }
+    auto whatFile = loc.file();
+    if (whatFile.exists()) {
+        if (auto doc = realmain::lsp::findDocumentation(whatFile.data(gs).source(), loc.beginPos())) {
+            docs.push_back(doc.value());
+        }
+    }
+}
+
+core::Loc GenericSymbolRef::symbolLoc(const core::GlobalState &gs) const {
+    switch (this->kind()) {
+        case Kind::Method: {
+            auto method = this->selfOrOwner.asMethodRef().data(gs);
+            if (!method->nameLoc.exists() || method->nameLoc.empty()) {
+                return method->loc();
+            }
+            return method->nameLoc;
+        }
+        case Kind::ClassOrModule:
+        case Kind::DeclaredField:
+            return this->selfOrOwner.loc(gs);
+        case Kind::UndeclaredField:
+            ENFORCE(false, "case UndeclaredField should not be triggered here");
+            return core::Loc();
+    }
+}
+
 } // namespace sorbet::scip_indexer
