+++ scip_indexer/SCIPSymbolRef.h
@@ -12,12 +12,16 @@
 #include "core/Loc.h"
 #include "core/NameRef.h"
 #include "core/SymbolRef.h"
+#include "core/Symbols.h"
+#include "core/TypePtr.h"
 
 namespace scip { // Avoid needlessly including protobuf header here.
 class Symbol;
 }
 
 namespace sorbet::scip_indexer {
+
+template <typename T> using SmallVec = InlinedVector<T, 2>;
 
 class GemMetadata final {
     std::string _name;
@@ -79,6 +83,148 @@ public:
     // Try to compute a scip::Symbol for this value.
     absl::Status symbolForExpr(const core::GlobalState &gs, const GemMetadata &metadata, std::optional<core::Loc> loc,
                                scip::Symbol &symbol) const;
+};
+
+// A wrapper type to handle both top-level symbols (like classes) as well as
+// "inner symbols" like fields (@x). In a statically typed language, field
+// symbols are like any other symbols, but in Ruby, they aren't (necessarily)
+// declared ahead-of-time (you can declare them with @x = T.let(…, …) though).
+// So Sorbet represents them with a separate name on the side.
+//
+// Structurally, this is similar to the Alias instruction. One key difference
+// is that a GenericSymbolRef may refer to the owner in some situations.
+class GenericSymbolRef final {
+    core::SymbolRef selfOrOwner;
+
+    /// Name of the symbol, which may or may not exist.
+    core::NameRef name;
+
+    /// The type of the symbol at its definition, if applicable.
+    ///
+    /// References to this symbol may have a different type,
+    /// because you can change the type of a field, including within
+    /// the same basic block.
+    core::TypePtr _definitionType;
+
+public:
+    enum class Kind {
+        ClassOrModule,
+        UndeclaredField,
+        DeclaredField,
+        Method,
+    };
+
+private:
+    GenericSymbolRef(core::SymbolRef s, core::NameRef n, core::TypePtr t, Kind k)
+        : selfOrOwner(s), name(n), _definitionType(t) {
+        switch (k) {
+            case Kind::ClassOrModule:
+                ENFORCE(s.isClassOrModule());
+                ENFORCE(!n.exists());
+                return;
+            case Kind::DeclaredField:
+                ENFORCE(s.isFieldOrStaticField());
+                ENFORCE(!n.exists());
+                return;
+            case Kind::UndeclaredField:
+                ENFORCE(s.isClassOrModule());
+                ENFORCE(n.exists());
+                return;
+            case Kind::Method:
+                ENFORCE(s.isMethod());
+                ENFORCE(!n.exists());
+        }
+    }
+
+public:
+    GenericSymbolRef(const GenericSymbolRef &) = default;
+    GenericSymbolRef(GenericSymbolRef &&) = default;
+    GenericSymbolRef &operator=(const GenericSymbolRef &) = default;
+    GenericSymbolRef &operator=(GenericSymbolRef &&) = default;
+
+    friend bool operator==(const GenericSymbolRef &lhs, const GenericSymbolRef &rhs) {
+        return lhs.selfOrOwner == rhs.selfOrOwner && lhs.name == rhs.name;
+    }
+
+    friend bool operator<(const GenericSymbolRef &lhs, const GenericSymbolRef &rhs) {
+        return lhs.selfOrOwner.rawId() < rhs.selfOrOwner.rawId() ||
+               (lhs.selfOrOwner == rhs.selfOrOwner && lhs.name.rawId() < rhs.name.rawId());
+    }
+
+    template <typename H> friend H AbslHashValue(H h, const GenericSymbolRef &c) {
+        return H::combine(std::move(h), c.selfOrOwner, c.name);
+    }
+
+    static GenericSymbolRef classOrModule(core::SymbolRef self) {
+        return GenericSymbolRef(self, {}, {}, Kind::ClassOrModule);
+    }
+
+    static GenericSymbolRef undeclaredField(core::SymbolRef owner, core::NameRef name, core::TypePtr type) {
+        return GenericSymbolRef(owner, name, type, Kind::UndeclaredField);
+    }
+
+    static GenericSymbolRef declaredField(core::SymbolRef self, core::TypePtr type) {
+        return GenericSymbolRef(self, {}, type, Kind::DeclaredField);
+    }
+
+    static GenericSymbolRef method(core::SymbolRef self) {
+        return GenericSymbolRef(self, {}, {}, Kind::Method);
+    }
+
+    core::TypePtr definitionType() const {
+        return this->_definitionType;
+    }
+
+    Kind kind() const {
+        if (this->name.exists()) {
+            return Kind::UndeclaredField;
+        }
+        if (this->selfOrOwner.isFieldOrStaticField()) {
+            return Kind::DeclaredField;
+        }
+        if (this->selfOrOwner.isMethod()) {
+            return Kind::Method;
+        }
+        return Kind::ClassOrModule;
+    }
+
+    UntypedGenericSymbolRef withoutType() const {
+        switch (this->kind()) {
+            case Kind::UndeclaredField:
+                ENFORCE(this->selfOrOwner.isClassOrModule());
+                return UntypedGenericSymbolRef::undeclared(this->selfOrOwner.asClassOrModuleRef(), this->name);
+            case Kind::Method:
+            case Kind::ClassOrModule:
+            case Kind::DeclaredField:
+                return UntypedGenericSymbolRef::declared(this->selfOrOwner);
+        }
+    }
+
+    /// Display a GenericSymbolRef for debugging.
+    std::string showRaw(const core::GlobalState &gs) const;
+
+    core::SymbolRef asSymbolRef() const {
+        ENFORCE(this->kind() != Kind::UndeclaredField);
+        return this->selfOrOwner;
+    }
+
+    static bool isSorbetInternal(const core::GlobalState &gs, core::SymbolRef sym);
+
+    bool isSorbetInternalClassOrMethod(const core::GlobalState &gs) const {
+        switch (this->kind()) {
+            case Kind::UndeclaredField:
+            case Kind::DeclaredField:
+                return false;
+            case Kind::ClassOrModule:
+            case Kind::Method:
+                return isSorbetInternal(gs, this->asSymbolRef());
+        }
+    }
+
+    void saveDocStrings(const core::GlobalState &gs, core::TypePtr fieldType, core::Loc loc,
+                        SmallVec<std::string> &docs) const;
+
+    core::Loc symbolLoc(const core::GlobalState &gs) const;
 };
 
 } // namespace sorbet::scip_indexer
